Updated EmailValidator and LoginHandler
Merged EMailvalidator into 1 file, in the LoginHandler.cs
