--- EmailValidator.cs
@@ -1,9 +1,0 @@
-﻿public class EmailValidator
-{
-    public bool IsValid(string email)
-    {
-        if (email.Length < 4 || !email.Contains("@"))
-            return false;
-        return true;
-    }
-}
